shell: fix signal handler id types
They are gulong, not guint. Avoids the warning:

warning: initialization of 'gulong * const' {aka 'long unsigned int * const'}
from incompatible pointer type 'guint *' {aka 'unsigned int *'}
[-Wincompatible-pointer-types]

Fixes 2f63071f968800df922f794793c436d76c663657

+++ shell/ev-recent-view.c
@@ -52,7 +52,7 @@ typedef struct {
         GtkListStore     *model;
         GtkRecentManager *recent_manager;
         GtkTreePath      *pressed_item_tree_path;
-        guint             recent_manager_changed_handler_id;
+	gulong            recent_manager_changed_handler_id;
 
 #ifdef HAVE_LIBGNOME_DESKTOP
         GnomeDesktopThumbnailFactory *thumbnail_factory;
